% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cali2010}
\alias{cali2010}
\title{Migration from California in 2010}
\format{
igraph object
}
\source{
\url{https://www.census.gov/data/tables/time-series/demo/geographic-mobility/state-to-state-migration.html}
}
\usage{
cali2010
}
\description{
A dataset containing the number of people who migrated from California to other US states
}
\keyword{datasets}
