% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_bundle.R
\name{edge_bundle}
\alias{edge_bundle}
\title{Edge bundling}
\usage{
edge_bundle(
  object,
  xy,
  type = c("force", "divided", "stub", "path", "hammer", "mingle"),
  ...
)
}
\arguments{
\item{object}{a graph object (igraph/network/tbl_graph)}

\item{xy}{coordinates of vertices}

\item{type}{bundling algorithm: \code{"force"} (force-directed, Holten),
\code{"divided"} (force-directed for directed graphs, Selassie et al.),
\code{"stub"} (Nocaj & Brandes), \code{"path"} (edge-path, Wallinger et al.),
\code{"hammer"} (KDE, Hurter et al.), or \code{"mingle"} (Gansner et al.)}

\item{...}{arguments passed on to the selected \verb{edge_bundle_*()} function}
}
\value{
data.frame containing the bundled edges (\code{x}, \code{y}, \code{index}, \code{group})
}
\description{
Dispatches to one of the edge bundling algorithms by name. This
is a thin wrapper over the individual \verb{edge_bundle_*()} functions and mirrors
the way frontends (e.g. vellumplot) select a bundling method.
}
\examples{
library(igraph)
g <- graph_from_edgelist(
    matrix(c(1, 12, 2, 11, 3, 10, 4, 9, 5, 8, 6, 7), ncol = 2, byrow = TRUE), FALSE
)
xy <- cbind(c(rep(0, 6), rep(1, 6)), c(1:6, 1:6))
edge_bundle(g, xy, type = "force")
}
\seealso{
\link{edge_bundle_force}, \link{edge_bundle_stub}, \link{edge_bundle_path},
\link{edge_bundle_hammer}, \link{edge_bundle_mingle}
}
\author{
David Schoch
}
