% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_path_biconn.R
\name{edge_bundle_path_biconn}
\alias{edge_bundle_path_biconn}
\title{edge_bundle_path_biconn
the function adds column \code{component}, which indicates which component the edge belongs to.}
\usage{
edge_bundle_path_biconn(
  graph,
  xy,
  weightFactor = 2,
  mode = "out",
  max_distortion = 2,
  segments = 20,
  bundle_strength = 1
)
}
\description{
edge_bundle_path_biconn
the function adds column \code{component}, which indicates which component the edge belongs to.
}
