% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgebundle-package.R
\docType{package}
\name{edgebundle-package}
\alias{edgebundle}
\alias{edgebundle-package}
\title{edgebundle: Algorithms for Bundling Edges in Networks and Visualizing Flow and Metro Maps}
\description{
An R package that implements several edge bundling/flow and metro map algorithms.
It includes:
\itemize{
\item Force directed edge bundling \code{edge_bundle_force()}
\item Stub bundling  \code{edge_bundle_stub()}
\item Hammer bundling  \code{edge_bundle_hammer()}
\item Edge-path bundling  \code{edge_bundle_path()}
\item TNSS flow map \verb{tnss_*()}
\item Multicriteria Metro map layout \code{metro_multicriteria()}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/edgebundle}
  \item \url{https://schochastics.github.io/edgebundle/}
  \item Report bugs at \url{https://github.com/schochastics/edgebundle/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

}
\keyword{internal}
