% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_hammer.R
\name{install_bundle_py}
\alias{install_bundle_py}
\title{install python dependencies for hammer bundling}
\usage{
install_bundle_py(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method (by default, "auto" automatically finds a
method that will work in the local environment, but note that the
"virtualenv" method is not available on Windows)}

\item{conda}{Path to conda executable (or "auto" to find conda using the PATH
and other conventional install locations)}
}
\description{
install datashader and scikit-image
}
