% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_flights}
\alias{us_flights}
\title{Flights within the US}
\format{
igraph object
}
\source{
\url{https://gist.githubusercontent.com/mbostock/7608400/raw}
}
\usage{
us_flights
}
\description{
A dataset containing flights between US airports as igraph object
}
\keyword{datasets}
