% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_migration}
\alias{us_migration}
\title{Migration within the US 2010-2019}
\format{
data.frame
}
\source{
\url{https://www.census.gov/data/tables/time-series/demo/geographic-mobility/state-to-state-migration.html}
}
\usage{
us_migration
}
\description{
A dataset containing the number of people migrating between US states from 2010-2019
}
\keyword{datasets}
